+++ NEWS.md
@@ -3,15 +3,16 @@
 * Added new dependency package `MGLM` to allow drawing random numbers from
 Dirichlet-multinomial distribution.
 
-* Added new argument `cp.factor` in function `AddNoise` to allow adding noise
-using Dirichlet-multinomial distribution.
+* Added new argument `cp.factor` in functions `AddNoise` and
+`GenerateNoisyTumors` to allow adding noise using Dirichlet-multinomial
+distribution.
 
 ### Changed
 * Changed the default value of `n.binom.size` to `NULL` in function `AddNoise`.
 
 <br> 
 
-## SynSigGen 1.0.13.9000
+## SynSigGen 1.0.14
 ### Fixed
 * Fixed a bug in `GenerateSyntheticExposures` when generating only one synthetic tumor
 with only one signature.